requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
minio==7.2.16 # 클라이언트
python-dotenv==1.1.1
pillow==11.3.0